LayoutTests/inspector-obsolete/http-tests/network/network-cyrillic-xhr-expected.txt

CONSOLE MESSAGE: line 6: XHR loaded: 1
CONSOLE MESSAGE: line 6: XHR loaded: 2
Tests cyrillic xhr content is correctly loaded in inspector.

Bug 79026
http://127.0.0.1:8000/inspector/network/resources/cyrillic.html
resource.content: The following line contains a word "success" written with cyrillic letters.
SUССЕSS

http://127.0.0.1:8000/inspector/network/resources/cyrillic.html
resource.content: The following line contains a word "success" written with cyrillic letters.
SUССЕSS
